Fixed one more uint32 usage

+++ lib/framebuffer.cc
@@ -439,7 +439,7 @@ Framebuffer::~Framebuffer() {
                                                   strlen("adafruit-hat")));
   // Initialize outputs, make sure that all of these are supported bits.
   const uint64_t result = io->InitOutputs(all_used_bits, is_some_adafruit_hat);
-  assert(result == (uint32_t)all_used_bits);  // Impl: all bits declared in gpio.cc ?
+  assert(result == all_used_bits);  // Impl: all bits declared in gpio.cc ?
 
   std::vector<int> bitplane_timings;
   uint32_t timing_ns = pwm_lsb_nanoseconds;
